+++ src/renderer/viz/expressions/RampImage.js
@@ -42,10 +42,10 @@ export default class RampImage extends Base {
         return paletteValues[index] || this.others.eval();
     }
 
-    getLegend (options) {
+    getLegendData (options) {
         const config = Object.assign({}, DEFAULT_OPTIONS, options);
         const type = this.input.type;
-        const legendData = this.input.getLegendData(config);
+        const legendData = this.input._getLegendData(config);
         const data = legendData.data.map(({key, value}) => {
             value = this._calcEval(value);
             return { key, value };